fix: invalid value separator
Invalid characters used as a value separator instead of a standard comma are ignored by both validator and schedule

For example: 1;2;3 * * * *

Validation passes (which is wrong), plus the scheduler represents it as "every first minute of every hour", instead of "every first, second and third minute of every hour".
the problem is in normalizeIntegers() that utilizes native JS function parseInt(), which behaves in unexpected ways.
If it is supplied with beforementioned string "1;2;3", it evaluates to 1.